requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
cfgv==3.4.0; python_version >= '3.8'
distlib==0.3.9
django==5.1.2; python_version >= '3.10'
django-debug-toolbar==4.4.6; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
djangorestframework==3.15.2; python_version >= '3.8'
environ==1.0
filelock==3.16.1; python_version >= '3.8'
identify==2.6.1; python_version >= '3.8'
nodeenv==1.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
platformdirs==4.3.6; python_version >= '3.8'
pre-commit==4.0.1; python_version >= '3.9'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.13'
virtualenv==20.27.1; python_version >= '3.8'
